educa/requirements.txt
asgiref==3.7.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.13
django-braces==1.15.0
django-debug-toolbar==4.3.0
django-embed-video==1.4.9
django-redisboard==8.4.0
djangorestframework==3.15.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==10.3.0
psycopg2-binary==2.9.10
pymemcache==4.0.0
python-decouple==3.8
redis==5.0.4
requests==2.31.0
setuptools==78.1.0
sqlparse==0.5.0
urllib3==2.3.0
uWSGI==2.0.28
